% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convertion_to_mod_percentage.R
\name{Convertion_to_mod_percentage}
\alias{Convertion_to_mod_percentage}
\title{Converts a flow value to a percentage of the module}
\usage{
Convertion_to_mod_percentage(Stream_Module, Flow)
}
\arguments{
\item{Stream_Module}{A numeric value corresponding to the average interannual flow rate ""module"".}

\item{Flow}{A numerical value corresponds to a flow rate which we wish to convert into a percentage of the average interannual flow rate "module".}
}
\value{
Numerical value of the "double" type which corresponds to the percentage flow rate of the module. Exemple: if Stream_Module=10 and Flow=16, then the flow rate in percentage of the module is 160\%
}
\description{
Converts a flow value to a percentage of the module
}
\examples{
Convertion_to_mod_percentage(Stream_Module=10,Flow=16)
}
