% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_zoom2.R
\name{facet_zoom2}
\alias{facet_zoom2}
\title{facet_zoom2}
\usage{
facet_zoom2(
  x,
  y,
  xy,
  zoom.data,
  xlim = NULL,
  ylim = NULL,
  split = FALSE,
  horizontal = TRUE,
  zoom.size = 2,
  show.area = TRUE,
  shrink = TRUE
)
}
\arguments{
\item{x}{An expression evaluating to a logical vector that determines the subset of data to zoom in on}

\item{y}{An expression evaluating to a logical vector that determines the subset of data to zoom in on}

\item{xy}{An expression evaluating to a logical vector that determines the subset of data to zoom in on}

\item{zoom.data}{An expression evaluating to a logical vector. If TRUE the data only shows in the zoom panels.
If FALSE the data only show in the context panel. If NA the data will show in all panels.}

\item{xlim}{Specific zoom ranges for each axis. If present they will override x, y, and/or xy.}

\item{ylim}{Specific zoom ranges for each axis. If present they will override x, y, and/or xy.}

\item{split}{If both x and y is given, should each axis zoom be shown separately as well? Defaults to FALSE}

\item{horizontal}{If both x and y is given and split = FALSE How should the zoom panel be positioned
relative to the full data panel? Defaults to TRUE}

\item{zoom.size}{Sets the relative size of the zoom panel to the full data panel. The default
(2) makes the zoom panel twice the size of the full data panel.}

\item{show.area}{Should the zoom area be drawn below the data points on the full data panel? Defaults to TRUE.}

\item{shrink}{If TRUE, will shrink scales to fit output of statistics, not raw data.
If FALSE, will be range of raw data before statistical summary.}
}
\value{
graphic
}
\description{
Alternative function to ggcorce::facet_zoom which allows to display a vertical zoom on the right side.
(and not on the left side as in the facet_zomm function)
}
