% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/flow_variation_considered_as_fast_as_possible_in_natural_hydrology.R
\name{flow_variation_considered_as_fast_as_possible_in_natural_hydrology}
\alias{flow_variation_considered_as_fast_as_possible_in_natural_hydrology}
\title{flow_variation_considered_as_fast_as_possible_in_natural_hydrology}
\usage{
flow_variation_considered_as_fast_as_possible_in_natural_hydrology(
  Stream_Module,
  b_drop = 0.75,
  b_rise = 1.15
)
}
\arguments{
\item{Stream_Module}{A numeric value corresponding to the average inter annual flow rate ""module"".}

\item{b_drop}{A numeric value of the coefficient \strong{b} for decrease. By default b_drop = 0.75}

\item{b_rise}{A numeric value of the coefficient \strong{b} for increase. By default b_rise = 1.15}
}
\value{
A table which reconstitutes the fastest variations in natural regime according to the module
}
\description{
\emph{Courret (2014)} shows that in natural hydrology the rise and fall times of hydrographs follow an exponential law.
\eqn{Q= 4 * exp(-a * t^b)}
The value of the coefficient \strong{a} can be expressed according to the module of the river.
The value of the coefficient \strong{b} is depends on the type of variation: increase or decrease
}
\examples{
flow_variation_considered_as_fast_as_possible_in_natural_hydrology(10)
}
